Change kyc migration

+++ src/main/resources/db/migration/V9__kyc_identyum.sql
@@ -1,2 +1,2 @@
 ALTER TABLE user_info ADD COLUMN identyum_user_uuid VARCHAR;
-ALTER TABLE coop ADD COLUMN kyc_provider_id SMALLINT NOT NULL DEFAULT 0;
+ALTER TABLE coop ADD COLUMN kyc_provider_id INT NOT NULL DEFAULT 0;
